+++ src/Util.cs
@@ -8,9 +8,9 @@ using TaleWorlds.Core.ViewModelCollection;
 
 namespace Pacemaker
 {
-	public class Util
+	class Util
 	{
-		public static bool EnableLog
+		internal static bool EnableLog
 		{
 			get
 			{
@@ -25,26 +25,26 @@ namespace Pacemaker
 			}
 		}
 
-		public static bool EnableTracer { get; set; } = false;
+		internal static bool EnableTracer { get; set; } = false;
 
-		public static GameLogBase Log = new GameLogBase(); // GameLogBase, parent of GameLog, implements do-nothing virtual output methods
+		internal static GameLogBase Log = new GameLogBase(); // GameLogBase, parent of GameLog, implements do-nothing virtual output methods
 
 		// Slightly out of place, but meh:
-		public static bool NearEqual(float v1, float v2, float epsilon = 1e-5f) => Math.Abs(v1 - v2) < epsilon;
-		public static bool NearEqual(double v1, double v2, double epsilon = 1e-5f) => Math.Abs(v1 - v2) < epsilon;
+		internal static bool NearEqual(float v1, float v2, float epsilon = 1e-5f) => Math.Abs(v1 - v2) < epsilon;
+		internal static bool NearEqual(double v1, double v2, double epsilon = 1e-5f) => Math.Abs(v1 - v2) < epsilon;
 
-		public static class EventTracer
+		internal static class EventTracer
 		{
 			private static readonly ConcurrentDictionary<string, bool> _stackTraceMap = new ConcurrentDictionary<string, bool>();
 
 			[MethodImpl(MethodImplOptions.NoInlining)]
-			public static void Trace(string extraInfo, int framesToSkip = 1)
+			internal static void Trace(string extraInfo, int framesToSkip = 1)
 			{
 				Trace(new List<string> { extraInfo }, framesToSkip + 1);
 			}
 
 			[MethodImpl(MethodImplOptions.NoInlining)]
-			public static void Trace(List<string> extraInfo = null, int framesToSkip = 1)
+			internal static void Trace(List<string> extraInfo = null, int framesToSkip = 1)
 			{
 				if (!EnableTracer || !EnableLog)
 					return;
@@ -66,12 +66,9 @@ namespace Pacemaker
 						$"Campaign Time:      {new SimpleTime(CampaignTime.Now)}",
 						$"  Elapsed Years:    {Campaign.Current.CampaignStartTime.ElapsedYearsUntilNow:F3}",
 						$"  Elapsed Seasons:  {Campaign.Current.CampaignStartTime.ElapsedSeasonsUntilNow:F3}",
-						$"  Elapsed Weeks:    {Campaign.Current.CampaignStartTime.ElapsedWeeksUntilNow:F2}",
 						$"  Elapsed Days:     {Campaign.Current.CampaignStartTime.ElapsedDaysUntilNow:F2}",
 						$"  Elapsed Hours:    {Campaign.Current.CampaignStartTime.ElapsedHoursUntilNow:F2}",
-						$"  Week of Season:   {CampaignTime.Now.GetWeekOfSeason + 1}",
 						$"  Day of Season:    {CampaignTime.Now.GetDayOfSeason + 1}",
-						$"  Day of Week:      {CampaignTime.Now.GetDayOfWeek + 1}",
 						$"  Hour of Day:      {CampaignTime.Now.GetHourOfDay}",
 					});
 				}